requirements.txt
openai>=1.14.0
python-dotenv>=1.0.0
requests>=2.31.0
tweepy>=4.14.0
facebook-sdk>=3.1.0
psycopg2-binary>=2.9.9
playwright>=1.42.0
